requirements.txt
torch
torchvision
torchaudio
transformers
datasets
evaluate
accelerate
librosa
soundfile
Pillow
pytesseract
numpy
bitsandbytes
gradio
python-dotenv